64-bit: Change page_index_t to 'int'
The max dynamic space size is contrained by the GC card size which uses
a (signed-byte 32) for the card index. Therefore cards of 1KiB can span
at most 2TiB. Assuming 32KiB pages, only 27 bits are needed to index
the highest numbered page, if my math is right. And mark-region-gc with
smaller cards could address at most 256GiB (8x less than gencgc).

+++ src/runtime/coreparse.c
@@ -1165,7 +1165,7 @@ void gc_load_corefile_ptes(int card_table_nbits,
 {
     if (next_free_page != n_ptes)
         lose("n_PTEs=%"PAGE_INDEX_FMT" but expected %"PAGE_INDEX_FMT,
-             n_ptes, next_free_page);
+             (int)n_ptes, next_free_page);
 
     gc_card_table_nbits = card_table_nbits;
     bool patchp = gc_allocate_ptes();
